+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 
 * ROOT: Disabled davix to resolve build issues
 * FairRoot: Add missing GEANT3 dependency
+* EvtGen: Fix detection of C++11
 
 ### Changed